requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.6
coloredlogs==15.0.1
crewai==0.75.0
crewai-tools==0.17.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
Django==5.1.4
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.126
executing==2.1.0
fastapi==0.115.6
fastavro==1.10.0
ffmpy==0.5.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.76.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
gradio==5.12.0
gradio_client==1.5.4
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.20
idna==3.10
image==1.5.33
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
instructor==1.6.3
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.5.0
jmespath==1.0.1
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.4
langchain-cohere==0.3.4
langchain-community==0.3.3
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.1.147
litellm==1.50.2
lomond==0.3.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.41
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.6
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.1.4
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.5
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.20.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.391
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.12.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.1
safehttpx==0.1.6
schema==0.7.7
selenium==4.27.1
semantic-version==2.10.0
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli_w==1.1.0
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
transformers>=4.30.0
torch>=2.0.0
pillow>=9.0.0
crewai>=0.14.0
python-dotenv>=1.0.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.5.16
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
